Restructuring grids submodule
